progress toward new setup of generating pages
